x509 Validation Support (#618)
* Initial x509 prototype, no unit tests, or stapled ocsp yet.

* Refactor timegit statusgit status

* Refactoringz. Made connection and config play nicely with validator. Still need to finish up the last bit of ocsp stapled validation.

* OCSP checks now verify the timestamps on the cert. updated unit tests, removed the verify_cb interface altogether.

* Start of x509 unit tests.

* Passing unit tests for cert chain validation. Still need to test ocsp and the host callback.

* Unit test that host name validation works and context is passed properly.

* Unit tests for OCSP, added some comments, and an ASN1 date parser. Moved around some reused code to testlib.

* Fixed some 'O Crap I forgot about that' stuff from looking at the PR. Nothing major.

* Addressed PR comments. Added more safety checks, cleaned up comments, made some variable names consistent.

* More PR updates. Cleaned up some comments and set error codes to safer defaults.

* Refactored clocks, updated s2nc and s2nd, reworked trust store to work with directories as well as ca files. Bug fixes after tests against live servers.

* Spare Colm the despair and self-loathing that would result from the c++ style comments he would need to call me out for. Removed some // style comments I noticed when reviewing the PR.

* Renamed clocks to monotonic and wall. Wrote some gates to handle boringssl not supporting ocsp validation.

* Added some tests for asn1 date parsing. Found a few bugs, fixed them. Trust store now can load locations multiple times, other msc. minor refactors.

* Secure is now the default. Also, trust store is automatically initialized to platform default paths. It's now an explicit call to disable client side validation. Unit test updates to handle new API.

* Updated usage guide with the x509 related API additions.

* Fixed compilation warnings in unit tests.

* Fixed some CI build (style failures), pulled asn1 time state machine into another function.

* Updates for Openssl 1.1.0, and fuzz test fixes.

* Attempt to make openssl 1.1.0 happy.

* Undo silly python 2.7 compat fix that broke python 3.

* added ca-certificates to ubuntu integration test run.

* Turned x509 validaiton off for travis integration test run.

* Removed openssl 1.1.0 compatibility fix from libressl code.

* Fixed typo in fuzz tests.

* Fixed stupid typos.

* Updates to fix integration test runs either to use a trust store, or disable validation.

* Added trust store to integration tests.

* Temporary strace and print of the commands being run for some hope of figuring out why trust store isn't working.

* Fixed typo in Travis config.

* Hail Mary.

* removed strace command, added current directory of python script as the working directory of the executable.

* Working directory should be directory not the actual filename.

* Fixed python arg passing that was breaking the test run.

* Ugh, another typo in python script.

* Added missing arguments for getopt_long for s2nd.

* got integration tests passing, by allowing partial chain if custom ca file is used.

* Test fixes. Added cleanup for static memory, updated libressl to later version (still needs to be uploaded to s3 bucket).

* additional commit to kick off CI.

* Appended libressl install path to bypass cache, memory leak fix, odd define issue on the build host wasn't picking up a symbol, added #define hack to work around it.

* Fixed idiotic read after free error.

* Fixed memory leak when set_config overwrites the connection's existing configuration.

* Addressed most CR comments. Updated libressl download path. The only part not done is the hostname validation which is still in discussion.

* Inverted auth, made hostname rejection the default for validation.

* Documentation update.

* Removed leftover flag.

* Addressed pr comments.

* moved some macros into s2n_openssl.h, after updating from master, used the S2N_ERROR_IF macro.

* Fixed line too long cppcheck error in s2n_server_cert.c

* Fixed branch statement spacing.

+++ tls/s2n_server_hello.c
@@ -45,7 +45,7 @@ int s2n_server_hello_recv(struct s2n_connection *conn)
 
     GUARD(s2n_stuffer_read_bytes(in, protocol_version, S2N_TLS_PROTOCOL_VERSION_LEN));
 
-    conn->server_protocol_version = (protocol_version[0] * 10) + protocol_version[1];
+    conn->server_protocol_version = (uint8_t)(protocol_version[0] * 10) + protocol_version[1];
 
     if (conn->server_protocol_version < conn->config->cipher_preferences->minimum_protocol_version || conn->server_protocol_version > conn->client_protocol_version) {
         GUARD(s2n_queue_reader_unsupported_protocol_version_alert(conn));
@@ -117,8 +117,8 @@ int s2n_server_hello_send(struct s2n_connection *conn)
     notnull_check(r.data);
     GUARD(s2n_get_public_random_data(&r));
 
-    protocol_version[0] = conn->actual_protocol_version / 10;
-    protocol_version[1] = conn->actual_protocol_version % 10;
+    protocol_version[0] = (uint8_t)(conn->actual_protocol_version / 10);
+    protocol_version[1] = (uint8_t)(conn->actual_protocol_version % 10);
 
 
     GUARD(s2n_stuffer_write_bytes(out, protocol_version, S2N_TLS_PROTOCOL_VERSION_LEN));
